feat: move Lighting Panel import and Light Vectors setup to RigOps
