+++ setup_wizard/geometry_nodes_setup/geometry_nodes_setups.py
@@ -1,12 +1,10 @@
 # Author: michael-gh1
 
 import bpy
-import os
 
 from abc import ABC, abstractmethod
 from bpy.types import Operator, Context
 
-from setup_wizard.geometry_nodes_setup.lighting_panel_names import LightingPanelNames
 from setup_wizard.domain.shader_node_names import StellarToonShaderNodeNames
 from setup_wizard.domain.shader_identifier_service import GenshinImpactShaders, HonkaiStarRailShaders, ShaderIdentifierService, ShaderIdentifierServiceFactory
 from setup_wizard.domain.shader_material_names import JaredNytsPunishingGrayRavenShaderMaterialNames, StellarToonShaderMaterialNames, V3_BonnyFestivityGenshinImpactMaterialNames, V2_FestivityGenshinImpactMaterialNames, ShaderMaterialNames, Nya222HonkaiStarRailShaderMaterialNames
@@ -233,28 +231,6 @@ class GameGeometryNodesSetup(ABC):
         light_vectors_modifier[LIGHT_VECTORS_HEAD_ORIGIN] = light_vectors_modifier[LIGHT_VECTORS_HEAD_ORIGIN] or bpy.data.objects.get(LightDirectionEmptyNames.HEAD_ORIGIN)
         light_vectors_modifier[LIGHT_VECTORS_HEAD_FORWARD] = light_vectors_modifier[LIGHT_VECTORS_HEAD_FORWARD] or bpy.data.objects.get(LightDirectionEmptyNames.HEAD_FORWARD)
         light_vectors_modifier[LIGHT_VECTORS_HEAD_UP] = light_vectors_modifier[LIGHT_VECTORS_HEAD_UP] or bpy.data.objects.get(LightDirectionEmptyNames.HEAD_UP) 
-
-        self.set_up_lighting_panel(light_vectors_modifier) 
-
-    # Genshin Shader >=v3.3
-    def set_up_lighting_panel(self, light_vectors_modifier):
-        if LightingPanelNames.LIGHT_VECTORS_MODIFIER_INPUT_NAME_TO_OBJECT_NAME[0][0] in light_vectors_modifier:
-            if not bpy.data.objects.get(LightingPanelNames.Objects.LIGHTING_PANEL):
-                self.import_lighting_panel()
-
-            for modifier_input_name, object_name in LightingPanelNames.LIGHT_VECTORS_MODIFIER_INPUT_NAME_TO_OBJECT_NAME:
-                light_vectors_modifier[modifier_input_name] = light_vectors_modifier[modifier_input_name] or bpy.data.objects.get(object_name)
-
-    def import_lighting_panel(self):
-        lighting_panel_filepath = os.path.join(os.path.dirname(os.path.abspath(__file__)), LightingPanelNames.FILENAME)
-        inner_path = 'Collection'
-        bpy.ops.wm.append(
-            filepath=os.path.join(lighting_panel_filepath, inner_path, LightingPanelNames.Collections.LIGHTING_PANEL),
-            directory=os.path.join(lighting_panel_filepath, inner_path),
-            files=[
-                {'name': LightingPanelNames.Collections.LIGHTING_PANEL},
-            ],
-        )
 
 
 class GenshinImpactGeometryNodesSetup(GameGeometryNodesSetup):
